requirements-lock.txt
absl-py==0.10.0
astunparse==1.6.3
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
contextlib2==21.6.0
cycler==0.10.0
decorator==5.0.9
dill==0.3.4
distro==1.5.0
dm-tree==0.1.6
flatbuffers==1.12
flax==0.3.1
future==0.18.2
gast==0.3.3
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
jax==0.2.8
jaxlib==0.1.60
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
ml-collections==0.1.0
msgpack==1.0.2
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.2.54
opt-einsum==3.3.0
packaging==20.9
pep517==0.10.0
Pillow==8.2.0
pip-tools==6.2.0
promise==2.3
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-build==0.11.1
scipy==1.7.0
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-addons==0.12.0
tensorflow-datasets==3.1.0
tensorflow-estimator==2.4.0
tensorflow-gan==2.0.0
tensorflow-hub==0.12.0
tensorflow-io==0.17.1
tensorflow-metadata==1.1.0
tensorflow-probability==0.12.2
termcolor==1.1.0
toml==0.10.2
torch==1.7.1
torchvision==0.8.2
tqdm==4.61.1
typeguard==2.12.1
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1